libgte: use `int*` instead of `long*` for params
This ensures the ported games will not have GTE function calls to write
4 extra unexpected bytes to the game structs on 64-bit builds.

+++ psyz/tests/test_libgte.cpp
@@ -28,9 +28,9 @@ class gte_Test : public testing::Test {
     struct RTPContext {
         MATRIX m = {0};
         SVECTOR svs[3] = {};
-        long lgs[3] = {};
-        long p = 0;
-        long flag = 0;
+        int lgs[3] = {};
+        int p = 0;
+        int flag = 0;
 
         RTPContext() {
             InitGeom();
